Moved module components out of DAR.
